requirements.txt
Babel==2.0
Jinja2==2.8
MarkupSafe==0.23
Pillow==2.9.0
Pygments==2.0.2
Sphinx==1.3.1
alabaster==0.7.6
docutils==0.12
ipython==3.2.1
matplotlib==1.4.3
nose==1.3.7
numpy==1.9.2
numpydoc==0.5
pandas==0.16.2
py==1.4.30
py-cpuinfo==0.1.6
-e git+git@github.com:ferdinandvwyk/pyfilm.git@f90a31dad697a1a22f93ea26f03d7fb21425487c#egg=pyfilm-origin/master
pyparsing==2.0.3
pytest==2.7.2
python-dateutil==2.4.2
pytz==2015.4
scipy==0.15.1
seaborn==0.6.0
six==1.9.0
snowballstemmer==1.2.0
sphinx-rtd-theme==0.1.8
